ENG-1327 Implement Conflict Management Strategy

+++ src/main/java/org/entando/kubernetes/model/bundle/installable/CategoryInstallable.java
@@ -3,6 +3,7 @@ package org.entando.kubernetes.model.bundle.installable;
 import java.util.concurrent.CompletableFuture;
 import lombok.extern.slf4j.Slf4j;
 import org.entando.kubernetes.client.core.EntandoCoreClient;
+import org.entando.kubernetes.controller.digitalexchange.job.model.InstallRequest.InstallAction;
 import org.entando.kubernetes.model.bundle.ComponentType;
 import org.entando.kubernetes.model.bundle.descriptor.CategoryDescriptor;
 
@@ -11,16 +12,26 @@ public class CategoryInstallable extends Installable<CategoryDescriptor> {
 
     private final EntandoCoreClient engineService;
 
-    public CategoryInstallable(EntandoCoreClient engineService, CategoryDescriptor descriptor) {
-        super(descriptor);
+    public CategoryInstallable(EntandoCoreClient engineService, CategoryDescriptor descriptor, InstallAction action) {
+        super(descriptor, action);
         this.engineService = engineService;
     }
 
     @Override
     public CompletableFuture<Void> install() {
         return CompletableFuture.runAsync(() -> {
-            log.info("Registering Label {}", getName());
-            engineService.registerCategory(representation);
+
+            logConflictStrategyAction();
+
+            if (shouldSkip()) {
+                return; //Do nothing
+            }
+
+            if (shouldCreate()) {
+                engineService.createCategory(representation);
+            } else {
+                engineService.updateCategory(representation);
+            }
         });
     }
 
@@ -28,7 +39,9 @@ public class CategoryInstallable extends Installable<CategoryDescriptor> {
     public CompletableFuture<Void> uninstall() {
         return CompletableFuture.runAsync(() -> {
             log.info("Removing Label {}", getName());
-            engineService.deleteCategory(getName());
+            if (shouldCreate()) {
+                engineService.deleteCategory(getName());
+            }
         });
     }
