+++ src/main/java/org/entando/kubernetes/model/bundle/installable/ContentTypeInstallable.java
@@ -3,6 +3,7 @@ package org.entando.kubernetes.model.bundle.installable;
 import java.util.concurrent.CompletableFuture;
 import lombok.extern.slf4j.Slf4j;
 import org.entando.kubernetes.client.core.EntandoCoreClient;
+import org.entando.kubernetes.controller.digitalexchange.job.model.InstallRequest.InstallAction;
 import org.entando.kubernetes.model.bundle.ComponentType;
 import org.entando.kubernetes.model.bundle.descriptor.contenttype.ContentTypeDescriptor;
 
@@ -11,16 +12,27 @@ public class ContentTypeInstallable extends Installable<ContentTypeDescriptor> {
 
     private final EntandoCoreClient engineService;
 
-    public ContentTypeInstallable(EntandoCoreClient service, ContentTypeDescriptor contentTypeDescriptor) {
-        super(contentTypeDescriptor);
+    public ContentTypeInstallable(EntandoCoreClient service, ContentTypeDescriptor contentTypeDescriptor,
+            InstallAction action) {
+        super(contentTypeDescriptor, action);
         this.engineService = service;
     }
 
     @Override
     public CompletableFuture<Void> install() {
         return CompletableFuture.runAsync(() -> {
-            log.info("Registering Content Type {}", getName());
-            engineService.registerContentType(representation);
+
+            logConflictStrategyAction();
+
+            if (shouldSkip()) {
+                return; //Do nothing
+            }
+
+            if (shouldCreate()) {
+                engineService.createContentType(representation);
+            } else {
+                engineService.updateContentType(representation);
+            }
         });
     }
 
@@ -28,7 +40,9 @@ public class ContentTypeInstallable extends Installable<ContentTypeDescriptor> {
     public CompletableFuture<Void> uninstall() {
         return CompletableFuture.runAsync(() -> {
             log.info("Removing Content Type {}", getName());
-            engineService.deleteContentType(getName());
+            if (shouldCreate()) {
+                engineService.deleteContentType(getName());
+            }
         });
     }
