+++ src/main/java/org/entando/kubernetes/model/bundle/installable/DirectoryInstallable.java
@@ -3,6 +3,7 @@ package org.entando.kubernetes.model.bundle.installable;
 import java.util.concurrent.CompletableFuture;
 import lombok.extern.slf4j.Slf4j;
 import org.entando.kubernetes.client.core.EntandoCoreClient;
+import org.entando.kubernetes.controller.digitalexchange.job.model.InstallRequest.InstallAction;
 import org.entando.kubernetes.model.bundle.ComponentType;
 import org.entando.kubernetes.model.bundle.descriptor.DirectoryDescriptor;
 
@@ -11,23 +12,22 @@ public class DirectoryInstallable extends Installable<DirectoryDescriptor> {
 
     private final EntandoCoreClient engineService;
 
-    public DirectoryInstallable(EntandoCoreClient engineService, DirectoryDescriptor directory) {
-        super(directory);
+    public DirectoryInstallable(EntandoCoreClient engineService, DirectoryDescriptor directory, InstallAction action) {
+        super(directory, action);
         this.engineService = engineService;
     }
 
     @Override
     public CompletableFuture<Void> install() {
         return CompletableFuture.runAsync(() -> {
-            log.info("Creating directory {}", getName());
-            engineService.createFolder(representation.getName());
+            //Do nothing
         });
     }
 
     @Override
     public CompletableFuture<Void> uninstall() {
         return CompletableFuture.runAsync(() -> {
-            if (this.representation.isRoot()) {
+            if (this.representation.isRoot() && shouldCreate()) {
                 engineService.deleteFolder(this.representation.getName());
             }
         });
