+++ src/main/java/org/entando/kubernetes/model/bundle/installable/FragmentInstallable.java
@@ -3,6 +3,7 @@ package org.entando.kubernetes.model.bundle.installable;
 import java.util.concurrent.CompletableFuture;
 import lombok.extern.slf4j.Slf4j;
 import org.entando.kubernetes.client.core.EntandoCoreClient;
+import org.entando.kubernetes.controller.digitalexchange.job.model.InstallRequest.InstallAction;
 import org.entando.kubernetes.model.bundle.ComponentType;
 import org.entando.kubernetes.model.bundle.descriptor.FragmentDescriptor;
 
@@ -11,16 +12,27 @@ public class FragmentInstallable extends Installable<FragmentDescriptor> {
 
     private final EntandoCoreClient engineService;
 
-    public FragmentInstallable(EntandoCoreClient engineService, FragmentDescriptor fragmentDescriptor) {
-        super(fragmentDescriptor);
+    public FragmentInstallable(EntandoCoreClient engineService, FragmentDescriptor fragmentDescriptor,
+            InstallAction action) {
+        super(fragmentDescriptor, action);
         this.engineService = engineService;
     }
 
     @Override
     public CompletableFuture<Void> install() {
         return CompletableFuture.runAsync(() -> {
-            log.info("Registering Fragment {}", getName());
-            engineService.registerFragment(representation);
+
+            logConflictStrategyAction();
+
+            if (shouldSkip()) {
+                return; //Do nothing
+            }
+
+            if (shouldCreate()) {
+                engineService.createFragment(representation);
+            } else {
+                engineService.updateFragment(representation);
+            }
         });
     }
 
@@ -28,7 +40,9 @@ public class FragmentInstallable extends Installable<FragmentDescriptor> {
     public CompletableFuture<Void> uninstall() {
         return CompletableFuture.runAsync(() -> {
             log.info("Removing Fragment {}", getName());
-            engineService.deleteFragment(getName());
+            if (shouldCreate()) {
+                engineService.deleteFragment(getName());
+            }
         });
     }
