+++ src/main/java/org/entando/kubernetes/model/bundle/installable/PageTemplateInstallable.java
@@ -3,6 +3,7 @@ package org.entando.kubernetes.model.bundle.installable;
 import java.util.concurrent.CompletableFuture;
 import lombok.extern.slf4j.Slf4j;
 import org.entando.kubernetes.client.core.EntandoCoreClient;
+import org.entando.kubernetes.controller.digitalexchange.job.model.InstallRequest.InstallAction;
 import org.entando.kubernetes.model.bundle.ComponentType;
 import org.entando.kubernetes.model.bundle.descriptor.PageTemplateDescriptor;
 
@@ -11,16 +12,27 @@ public class PageTemplateInstallable extends Installable<PageTemplateDescriptor>
 
     private final EntandoCoreClient engineService;
 
-    public PageTemplateInstallable(EntandoCoreClient engineService, PageTemplateDescriptor pageTemplateDescriptor) {
-        super(pageTemplateDescriptor);
+    public PageTemplateInstallable(EntandoCoreClient engineService, PageTemplateDescriptor pageTemplateDescriptor,
+            InstallAction action) {
+        super(pageTemplateDescriptor, action);
         this.engineService = engineService;
     }
 
     @Override
     public CompletableFuture<Void> install() {
         return CompletableFuture.runAsync(() -> {
-            log.info("Registering Page Model {}", getName());
-            engineService.registerPageModel(representation);
+
+            logConflictStrategyAction();
+
+            if (shouldSkip()) {
+                return; //Do nothing
+            }
+
+            if (shouldCreate()) {
+                engineService.createPageTemplate(representation);
+            } else {
+                engineService.updatePageTemplate(representation);
+            }
         });
     }
 
@@ -28,7 +40,9 @@ public class PageTemplateInstallable extends Installable<PageTemplateDescriptor>
     public CompletableFuture<Void> uninstall() {
         return CompletableFuture.runAsync(() -> {
             log.info("Removing PageTemplate {}", getName());
-            engineService.deletePageModel(getName());
+            if (shouldCreate()) {
+                engineService.deletePageModel(getName());
+            }
         });
     }
