+++ src/main/java/org/entando/kubernetes/model/bundle/installable/WidgetInstallable.java
@@ -3,6 +3,7 @@ package org.entando.kubernetes.model.bundle.installable;
 import java.util.concurrent.CompletableFuture;
 import lombok.extern.slf4j.Slf4j;
 import org.entando.kubernetes.client.core.EntandoCoreClient;
+import org.entando.kubernetes.controller.digitalexchange.job.model.InstallRequest.InstallAction;
 import org.entando.kubernetes.model.bundle.ComponentType;
 import org.entando.kubernetes.model.bundle.descriptor.WidgetDescriptor;
 
@@ -11,16 +12,26 @@ public class WidgetInstallable extends Installable<WidgetDescriptor> {
 
     private final EntandoCoreClient engineService;
 
-    public WidgetInstallable(EntandoCoreClient engineService, WidgetDescriptor widgetDescriptor) {
-        super(widgetDescriptor);
+    public WidgetInstallable(EntandoCoreClient engineService, WidgetDescriptor widgetDescriptor, InstallAction action) {
+        super(widgetDescriptor, action);
         this.engineService = engineService;
     }
 
     @Override
     public CompletableFuture<Void> install() {
         return CompletableFuture.runAsync(() -> {
-            log.info("Registering Widget {}", getName());
-            engineService.registerWidget(representation);
+
+            logConflictStrategyAction();
+
+            if (shouldSkip()) {
+                return; //Do nothing
+            }
+
+            if (shouldCreate()) {
+                engineService.createWidget(representation);
+            } else {
+                engineService.updateWidget(representation);
+            }
         });
     }
 
@@ -28,7 +39,9 @@ public class WidgetInstallable extends Installable<WidgetDescriptor> {
     public CompletableFuture<Void> uninstall() {
         return CompletableFuture.runAsync(() -> {
             log.info("Removing Widget {}", getName());
-            engineService.deleteWidget(getName());
+            if (shouldCreate()) {
+                engineService.deleteWidget(getName());
+            }
         });
     }
